[Style][Utils] Move device_busy_of beside the device properties
The helper times any callable on any device and carries nothing
reduction-specific; tileops.utils is where the device-level utilities
live. tune_by_forward stays in _primitives: it reaches into a kernel's
config candidates and is family machinery.
