+++ src/tileops/utils/__init__.py
@@ -1,5 +1,6 @@
 from .utils import (
     WARP_LANES,
+    device_busy_of,
     forget_device_properties,
     get_sm_count,
     get_sm_version,
@@ -9,6 +10,7 @@ from .utils import (
 
 __all__ = [
     "WARP_LANES",
+    "device_busy_of",
     "forget_device_properties",
     "get_sm_count",
     "get_sm_version",
