+++ src/tileops/utils/utils.py
@@ -64,3 +64,34 @@ def forget_device_properties() -> None:
     """
     _device_name.cache_clear()
     _sm_version.cache_clear()
+
+
+# Spin cycles queued before a device_busy_of measurement: tens of milliseconds
+# on any supported clock, ample to enqueue every timed call first.
+_BUSY_TIMING_SPIN_CYCLES = 50_000_000
+
+
+def device_busy_of(call, device: "torch.device", warmup: int = 5, rep: int = 20) -> float:
+    """Mean device time of *call* in milliseconds with host gaps excluded.
+
+    Judges paths that launch different kernel counts by their GPU work alone;
+    wall latency would charge a multi-launch path the host gaps between its
+    launches. A spin kernel holds the device while every timed call is
+    enqueued, so the queue then drains back to back and the event pair brackets
+    execution only. Deliberately not a profiler: the benchmark's own collector
+    owns the process's CUPTI subscription, and a second subscriber would break
+    its kernel attribution for the rest of the process.
+    """
+    with torch.cuda.device(device):
+        for _ in range(warmup):
+            call()
+        torch.cuda.synchronize()
+        start = torch.cuda.Event(enable_timing=True)
+        end = torch.cuda.Event(enable_timing=True)
+        torch.cuda._sleep(_BUSY_TIMING_SPIN_CYCLES)
+        start.record()
+        for _ in range(rep):
+            call()
+        end.record()
+        torch.cuda.synchronize()
+        return start.elapsed_time(end) / rep
